Revert "Start moving runtime types required by Connection to their own file."
