+++ packages/malloy/src/lang/ast/field-space/reference-field.ts
@@ -22,6 +22,7 @@
  */
 
 import {
+  Annotation,
   FieldAtomicDef,
   QueryFieldDef,
   TypeDesc,
@@ -29,7 +30,6 @@ import {
   isFilteredAliasedName,
   hasExpression,
 } from '../../../model/malloy_types';
-import {Annotation} from '../../../model/model_runtime_types';
 
 import {FieldReference} from '../query-items/field-references';
 import {FieldSpace} from '../types/field-space';
