+++ packages/malloy/src/lang/ast/field-space/rename-space-field.ts
@@ -21,8 +21,11 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import {FieldDef, TypeDesc} from '../../../model/malloy_types';
-import {DocumentLocation} from '../../../model/model_runtime_types';
+import {
+  DocumentLocation,
+  FieldDef,
+  TypeDesc,
+} from '../../../model/malloy_types';
 
 import {SpaceField} from '../types/space-field';
