+++ packages/malloy/src/lang/ast/query-properties/nest.ts
@@ -22,7 +22,6 @@
  */
 
 import * as model from '../../../model/malloy_types';
-import {Annotation} from '../../../model/model_runtime_types';
 import {FieldSpace} from '../types/field-space';
 import {MalloyElement} from '../types/malloy-element';
 import {Noteable, extendNoteMethod} from '../types/noteable';
@@ -66,7 +65,7 @@ abstract class TurtleDeclRoot
 {
   readonly isNoteableObj = true;
   extendNote = extendNoteMethod;
-  note?: Annotation;
+  note?: model.Annotation;
   constructor(readonly name: string) {
     super();
   }
