+++ packages/malloy/src/lang/malloy-to-ast.ts
@@ -43,8 +43,12 @@ import {
   idToStr,
 } from './parse-utils';
 import {CastType} from '../model';
-import {isCastType, isMatrixOperation} from '../model/malloy_types';
-import {DocumentLocation, Note} from '../model/model_runtime_types';
+import {
+  DocumentLocation,
+  isCastType,
+  isMatrixOperation,
+  Note,
+} from '../model/malloy_types';
 import {Tag} from '../tags';
 
 class ErrorNode extends ast.SourceQueryNode {
