+++ packages/malloy/src/lang/parse-log.ts
@@ -21,7 +21,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import {DocumentLocation} from '../model/model_runtime_types';
+import {DocumentLocation} from '../model/malloy_types';
 
 export type LogSeverity = 'error' | 'warn' | 'debug';
