+++ packages/malloy/src/lang/parse-tree-walkers/model-annotation-walker.ts
@@ -26,11 +26,7 @@ import {ParseTreeWalker} from 'antlr4ts/tree/ParseTreeWalker';
 import {MalloyParserListener} from '../lib/Malloy/MalloyParserListener';
 import * as parser from '../lib/Malloy/MalloyParser';
 import {MalloyTranslation} from '../parse-malloy';
-import {
-  Annotation,
-  DocumentLocation,
-  Note,
-} from '../../model/model_runtime_types';
+import {Annotation, DocumentLocation, Note} from '../../model/malloy_types';
 import {MalloyParseInfo} from '../malloy-parse-info';
 
 class ModelAnnotationWalker implements MalloyParserListener {
