+++ packages/malloy/src/lang/test/locations.spec.ts
@@ -33,10 +33,7 @@ import {
   model,
 } from './test-translator';
 import './parse-expects';
-import {
-  DocumentLocation,
-  DocumentPosition,
-} from '../../model/model_runtime_types';
+import {DocumentLocation, DocumentPosition} from '../../model/malloy_types';
 
 describe('source locations', () => {
   test('renamed source location', () => {
