+++ packages/malloy/src/model/malloy_types.ts
@@ -23,18 +23,6 @@
 
 // clang-format off
 
-import {
-  SQLBlock,
-  QueryRunStats,
-  DocumentLocation,
-  Annotation,
-  NamedObject,
-  AliasedName,
-  QueryData,
-  HasLocation,
-  QueryValue,
-} from './model_runtime_types';
-
 interface ParamBase {
   name: string;
   type: AtomicFieldType;
@@ -60,9 +48,24 @@ export function paramHasValue(p: Parameter): boolean {
   return isValueParameter(p) || p.condition !== null;
 }
 
+export interface DocumentRange {
+  start: DocumentPosition;
+  end: DocumentPosition;
+}
+
+export interface DocumentPosition {
+  line: number;
+  character: number;
+}
+
 export interface ImportLocation {
   importURL: string;
   location: DocumentLocation;
+}
+
+export interface DocumentLocation {
+  url: string;
+  range: DocumentRange;
 }
 
 interface DocumentReferenceBase {
@@ -103,6 +106,17 @@ export type DocumentReference =
   | DocumentFieldReference
   | DocumentJoinReference;
 
+/** put location into the parse tree. */
+export interface HasLocation {
+  location?: DocumentLocation;
+}
+
+/** All names have their source names and how they will appear in the symbol table that owns them */
+export interface AliasedName {
+  name: string;
+  as?: string;
+}
+
 export interface TypedObject {
   type: string;
 }
@@ -119,6 +133,11 @@ export function isFilteredAliasedName(
     }
   }
   return true;
+}
+
+/** all named objects have a type an a name (optionally aliased) */
+export interface NamedObject extends AliasedName, HasLocation {
+  type: string;
 }
 
 // result metadata for a field
@@ -871,6 +890,12 @@ export interface SQLBlockSource {
   select: SQLPhrase[];
 }
 
+export interface SQLBlock extends NamedObject {
+  type: 'sqlBlock';
+  connection?: string;
+  selectStr: string;
+}
+
 interface SubquerySource {
   type: 'sql';
   method: 'subquery';
@@ -1091,10 +1116,44 @@ export interface ModelDef {
 export type NamedStructDefs = Record<string, StructDef>;
 export type NamedModelObjects = Record<string, NamedModelObject>;
 
+/** Malloy source annotations attached to objects */
+export interface Annotation {
+  inherits?: Annotation;
+  blockNotes?: Note[];
+  notes?: Note[];
+}
+export interface Note {
+  text: string;
+  at: DocumentLocation;
+}
 /** Annotations with a uuid to make it easier to stream */
 export interface ModelAnnotation extends Annotation {
   id: string;
 }
+
+export type QueryScalar = string | boolean | number | Date | Buffer | null;
+
+/** One value in one column of returned data. */
+export type QueryValue = QueryScalar | QueryData | QueryDataRow;
+
+/** A row of returned data. */
+export type QueryDataRow = {[columnName: string]: QueryValue};
+
+/** Returned query data. */
+export type QueryData = QueryDataRow[];
+
+/** Query execution stats. */
+export type QueryRunStats = {
+  queryCostBytes?: number;
+};
+
+/** Returned Malloy query data */
+export type MalloyQueryData = {
+  rows: QueryDataRow[];
+  totalRows: number;
+  runStats?: QueryRunStats;
+  profilingUrl?: string;
+};
 
 export interface DrillSource {
   sourceExplore: string;
@@ -1218,3 +1277,4 @@ export interface SearchValueMapResult {
   cardinality: number;
   values: {fieldValue: string; weight: number}[];
 }
+// clang-format on
